+++ home-automation-ui/src/App.js
@@ -3,6 +3,7 @@ import { Route, Switch } from "react-router-dom";
 
 import "./App.module.scss";
 import Layout from "./hoc/Layout/Layout";
+import Guard from "./hoc/Guard/Guard";
 import RoomsDashboard from "./containers/RoomsDashboard/RoomsDashboard";
 import asyncComponent from "./hoc/asyncComponent/asyncComponent";
 
@@ -14,8 +15,10 @@ function App() {
   return (
     <Layout>
       <Switch>
-        <Route path="/room/:id" exact component={AsyncRoomsDevices} />
-        <Route path="/" exact component={RoomsDashboard} />
+        <Guard>
+          <Route path="/" exact component={RoomsDashboard} />
+          <Route path="/room/:id" exact component={AsyncRoomsDevices} />
+        </Guard>
       </Switch>
     </Layout>
   );