+++ home-automation-ui/src/store/auth/auth.actiontypes.js
@@ -1,0 +1,5 @@
+export const REGISTER_SUCCESS = "[AUTH] REGISTER_SUCCESS";
+export const REGISTER_FAILED = "[AUTH] REGISTER_FAILED";
+export const LOGIN_SUCCESS = "[AUTH] LOGIN_SUCCESS";
+export const LOGIN_FAILED = "[AUTH] LOGIN_FAILED";
+export const LOGOUT = "[AUTH] LOGOUT";